On branch main  Changes to be committed: 	modified:   frontend/src/components/technical_service/HomeTechnicalService.jsx 	modified:   frontend/src/components/technical_service/SearchedWarrantyDetailsModal.jsx 	modified:   frontend/src/components/technical_service/WarrantyDetailsModal.jsx
Modified behavior in all the Technical Services buttons that set loading to false in the catch block of their try/catch statements. Now, if an error occurs, the loading state will be set to false, preventing the loading spinner from persisting indefinitely. This ensures a better user experience by allowing users to see that the operation has completed, even if it was unsuccessful.